Add authorization handler for users controller;

+++ MathEvent.Server/MathEvent.Api/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using MathEvent.AuthorizationHandlers;
 using MathEvent.Converters.Identities.DTOs;
 using MathEvent.Converters.Identities.Models;
 using MathEvent.Converters.Others;
@@ -19,10 +20,16 @@ namespace MathEvent.Api.Controllers
 
         private readonly UserService _userService;
 
-        public UsersController(IMapper mapper, UserService userService)
+        private readonly IAuthorizationService _authorizationService;
+
+        public UsersController(
+            IMapper mapper,
+            UserService userService,
+            IAuthorizationService authorizationService)
         {
             _mapper = mapper;
             _userService = userService;
+            _authorizationService = authorizationService;
         }
 
         // GET api/Users/?key1=value1&key2=value2
@@ -100,6 +107,14 @@ namespace MathEvent.Api.Controllers
                 return NotFound(userResult.Messages);
             }
 
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, userResult.Entity, Operations.Update);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return StatusCode(403);
+            }
+
             var updateResult = await _userService.UpdateAsync(id, userUpdateModel);
 
             if (updateResult.Succeeded)
@@ -139,6 +154,14 @@ namespace MathEvent.Api.Controllers
             {
                 return NotFound(userResult.Messages);
             }
+
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, userResult.Entity, Operations.Update);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return StatusCode(403);
+            };
 
             var userEntity = userResult.Entity;
             var userDTO = _mapper.Map<UserWithEventsDTO>(userEntity);
@@ -183,6 +206,14 @@ namespace MathEvent.Api.Controllers
             if (!userResult.Succeeded)
             {
                 return NotFound(userResult.Messages);
+            }
+
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, userResult.Entity, Operations.Delete);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return StatusCode(403);
             }
 
             var deleteResult = await _userService.DeleteAsync(id);
